refactor code and move addTask to loader file
